move suite config to suite package
