+++ internal/cypress/v1alpha/config.go
@@ -12,6 +12,7 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/saucelabs/saucectl/internal/concurrency"
 	"github.com/saucelabs/saucectl/internal/config"
+	"github.com/saucelabs/saucectl/internal/cypress/suite"
 	"github.com/saucelabs/saucectl/internal/fpath"
 	"github.com/saucelabs/saucectl/internal/msg"
 	"github.com/saucelabs/saucectl/internal/region"
@@ -487,10 +488,10 @@ func (p *Project) SetCLIFlags(flags map[string]interface{}) {
 }
 
 // GetSuites returns suites
-func (p *Project) GetSuites() []config.Suite {
-	suites := []config.Suite{}
+func (p *Project) GetSuites() []suite.Suite {
+	suites := []suite.Suite{}
 	for _, s := range p.Suites {
-		suites = append(suites, config.Suite{
+		suites = append(suites, suite.Suite{
 			Name:             s.Name,
 			Browser:          s.Browser,
 			BrowserVersion:   s.BrowserVersion,
@@ -514,9 +515,9 @@ func (p *Project) GetKind() string {
 }
 
 // GetSuite returns suite
-func (p *Project) GetSuite() config.Suite {
+func (p *Project) GetSuite() suite.Suite {
 	s := p.Suite
-	return config.Suite{
+	return suite.Suite{
 		Name:             s.Name,
 		Browser:          s.Browser,
 		BrowserVersion:   s.BrowserVersion,
